+++ INF404/TP4/ast_parcours.h
@@ -4,7 +4,7 @@
 void afficherAst(Ast expr) ;
 // affiche l'arbre abstrait de l'expression arithmetique expr
 
-int evaluation(Ast expr) ;
+double evaluation(Ast expr) ;
 // calcule la valeur de l'expression arithmetique expr
 // FONCTION A COMPLETER !
